linebot-ai/requirements.txt
line-bot-sdk>=3.5.0
python-dotenv>=1.0.0
flask>=2.0.0
gunicorn>=20.1.0 